use int64_t to store score

+++ be/src/storage/compaction_candidate.h
@@ -60,8 +60,8 @@ struct CompactionCandidate {
 // when compaction score and level are equal, use tablet id(to be unique) instead(ascending)
 struct CompactionCandidateComparator {
     bool operator()(const CompactionCandidate& left, const CompactionCandidate& right) const {
-        int32_t left_score = static_cast<int32_t>(left.tablet->compaction_score(left.level) * 100);
-        int32_t right_score = static_cast<int32_t>(right.tablet->compaction_score(right.level) * 100);
+        int64_t left_score = static_cast<int64_t>(left.tablet->compaction_score(left.level) * 100);
+        int64_t right_score = static_cast<int64_t>(right.tablet->compaction_score(right.level) * 100);
         return left_score > right_score ||
                (left_score == right_score && left.level < right.level) ||
                (left_score == right_score && left.level == right.level &&
